Add files via upload
The goal of this project is to predict salary of an employee on the basis of their job experience. This project can take the database for the salary system from the organization and makes a graph through this information from the database. It will check the salary fields then import a graph which helps to observe the graphical representation. And then it can predict a salary of any employee through the prediction algorithm. It can also be applied in some other effective prediction also.